requirements.txt
anyio==3.6.2
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.5.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
ccsm==0.9.14.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
ecdsa==0.18.0
eyeD3==0.8.10
fastapi==0.89.1
filelock==3.6.0
greenlet==2.0.2
grpcio==1.30.2
h11==0.14.0
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
Mako==1.1.3
MarkupSafe==2.1.2
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.6.2
protobuf==3.12.4
psutil==5.9.0
psycopg==3.1.8
psycopg2==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pydantic==1.10.4
pyelftools==0.27
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
python-apt==2.4.0
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
python-jose==3.3.0
python-magic==0.4.24
python-multipart==0.0.5
python-xlib==0.29
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
rsa==4.9
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.1
starlette==0.22.0
systemd-python==234
tldextract==3.1.2
typing_extensions==4.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
uvicorn==0.20.0
virtualenv==20.17.1
wadllib==1.3.6
Werkzeug==2.2.2
wsproto==1.0.0
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
